refactor: extract sensor logic into internal classes, use coroutines instead of Handler
- Extract SensorManager from MainActivity into separate internal class
- Extract TouchSensorManager from TouchActivity into separate internal class
- Replace string literals with companion object constants
- Replace Handler postDelayed with Kotlin Coroutines (launch + delay)
